requirements.txt
matplotlib
pandas
plotly
python-dotenv